requirements.txt
Django==1.7.7
MySQL-python==1.2.5
argparse==1.2.1
django-configurations==0.8
wsgiref==0.1.2
django-bootstrap3
django-form-utils
django-widget-tweaks